Refactored the code to do the changes requested close #7
